feat: base collector upgrade, same interface as bribe collector (#637)
# `HarvestBaseCollector` upgrade

Motivation behind the upgrade is to make a uniform interface between
`BribeCollector` and base collector for simplicity of auction bot.
Specifically this involves:
- adding `payoutToken` as a constant, as it will always be wbera
- adding `claimFees` which also enables tokens other than iBGT to be
auctioned for wbera (eg IR in future)

## Low risk upgrade

HarvestBaseCollector has yet to be used, has no specific dependencies
and the upgrade inherits prior contract with no storage changes.